Объединил модели, составляющие сущность паспорта в одну модель Passport. Добавил модель(EvalMark) метки качества и наличия проведения оценок соблюдения каких-либо государственных стандартов. Например, оценка СОУТ или оценка профрисков в организации
