+++ app.py
@@ -3,7 +3,7 @@ from flask import Flask
 from config import AppConfig
 from data.db import db_sessionmaker
 from data.routes.all_routes import auth
-from data.routes import routes_manager
+# from data.routes import routes_manager  # заглушено на время интеграции с бд
 from flask import render_template
 
 
@@ -28,6 +28,7 @@ if __name__ == '__main__':
     @app.errorhandler(401)
     def page_not_found(e):
         return render_template('403.html'), 401
-
-    routes_manager.global_routes_init(app) 
+    
+    # заглушено на время интеграции с бд
+    # routes_manager.global_routes_init(app)
     app.run(host='127.0.0.1', port=8080, debug=True)
