+++ data/db/__all_models.py
@@ -1,19 +1,11 @@
-# Модели, представляющие сущность паспорта организации
-from .models.passport_entity.collective_agreement import CollectiveAgreement
-from .models.passport_entity.general_data import GeneralData
-from .models.passport_entity.golden_badge import GoldenBadge
-from .models.passport_entity.injuries import Injuries
-from .models.passport_entity.passport import Passport
-from .models.passport_entity.profrisk import Profrisk
-from .models.passport_entity.safety_training import SafetyTraining
-from .models.passport_entity.sout import Sout
-from .models.passport_entity.work_condition import WorkCondition
-# Модели, тоже относящиеся к системе паспортов, но
-# не имеющие ONE-TO-ONE связь с моделью Passport
+# Модели, представляющие собой систему паспортов
+from .models.passport_models.eval_mark import EvalMark
+from .models.passport_models.passport import Passport
 from .models.passport_models.passport_status import PassportStatus
 from .models.passport_models.photo import Photo
 from .models.passport_models.video import Video
-# Модели, отвечающие за опросы
+# Модели, отвечающие за работу универсальной формы опросника
+# для сбора информации об охране труда в организациях
 from .models.survey_models.answer import Answer
 from .models.survey_models.field import Field
 from .models.survey_models.field_type import FieldType
